basic validation

+++ src/book/book.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Book } from './schemas/book.schema';
 import * as mongoose from 'mongoose';
@@ -38,6 +42,12 @@ export class BookService {
   }
 
   async findById(id: string): Promise<Book> {
+    const isValidId = mongoose.isValidObjectId(id);
+
+    if (!isValidId) {
+      throw new BadRequestException('Please enter a valid id');
+    }
+
     const book = await this.bookModel.findById(id);
 
     if (!book) {
@@ -47,6 +57,12 @@ export class BookService {
   }
 
   async updateById(id: string, book: Book): Promise<Book> {
+    const isValidId = mongoose.isValidObjectId(id);
+
+    if (!isValidId) {
+      throw new BadRequestException('Please enter a valid id');
+    }
+
     return await this.bookModel.findByIdAndUpdate(id, book, {
       new: true,
       runValidators: true,
@@ -54,6 +70,12 @@ export class BookService {
   }
 
   async deleteById(id: string): Promise<Book> {
+    const isValidId = mongoose.isValidObjectId(id);
+
+    if (!isValidId) {
+      throw new BadRequestException('Please enter a valid id');
+    }
+
     const res = await this.bookModel.findByIdAndDelete(id, {
       new: true,
       runValidators: true,
